Fixed CodeMaid definition - enforcing rules so we don't use defaults. Folder refactoring - moved files around for better organization
