+++ source/CodeMaid.config
@@ -19,8 +19,7 @@
       <setting name="Reorganizing_RunAtStartOfCleanup" serializeAs="String">
         <value>True</value>
       </setting>
-      <setting name="Reorganizing_AlphabetizeMembersOfTheSameGroup"
-        serializeAs="String">
+      <setting name="Reorganizing_AlphabetizeMembersOfTheSameGroup" serializeAs="String">
         <value>False</value>
       </setting>
       <setting name="Reorganizing_ExplicitMembersAtEnd" serializeAs="String">
@@ -34,10 +33,10 @@
         <value>1</value>
       </setting>
       <setting name="Reorganizing_PrimaryOrderByAccessLevel" serializeAs="String">
-        <value>True</value>
+        <value>False</value>
       </setting>
       <setting name="Reorganizing_ReverseOrderByAccessLevel" serializeAs="String">
-        <value>False</value>
+        <value>True</value>
       </setting>
       <setting name="Digging_PrimarySortOrder" serializeAs="String">
         <value>1</value>
