+++ source/Xeno.ToolsHub.SampleAssmAddin/Handlers/PreferenceHandler.cs
@@ -1,29 +1,34 @@
 ﻿/* Copyright Xeno Innovations, Inc. 2018
- * Date:    2018-12-7
+ * Date:    2018-12-5
  * Author:  Damian Suess
  * File:    PreferenceHandler.cs
  * Description:
- *  Sample XML Add-in Preference Page handler
+ *  Sample Preference Page handler
  */
 
 using System.Windows.Forms;
+using Xeno.ToolsHub.ExtensionModel;
 using Xeno.ToolsHub.ExtensionModel.Preferences;
+using Xeno.ToolsHub.SampleAssmAddin.Views;
 
-namespace Xeno.ToolsHub.SampleXmlAddin
+namespace Xeno.ToolsHub.SampleAssmAddin.Handlers
 {
+  [Mono.Addins.Extension(
+    NodeName = ExtensionName.PreferencePageAddin,
+    Path = ExtensionPath.PreferencePage)]
   public class PreferenceHandler : PreferencePageExtension
   {
-    private PerferencePage _page;
+    private PreferencesPage _page;
 
     public override Form Page => _page;
 
-    public override string Title => "Sample Ext-XML";
+    public override string Title { get { return "Sample Ext-Assm"; } }
 
     public override bool IsModified => _page.IsModified;
 
     public override void InitializePage()
     {
-      _page = new PerferencePage();
+      _page = new PreferencesPage();
     }
 
     public override void OnSave()
