+++ source/Xeno.ToolsHub.SampleAssmAddin/Handlers/UtilityHandler.cs
@@ -9,14 +9,14 @@
 using Xeno.ToolsHub.Config;
 using Xeno.ToolsHub.ExtensionModel;
 
-namespace Xeno.ToolsHub.SampleAssmAddin
+namespace Xeno.ToolsHub.SampleAssmAddin.Handlers
 {
   [Mono.Addins.Extension(NodeName = ExtensionName.UtilityAddin, Path = ExtensionPath.Utility)]
-  public class AddinUtilityHandler : UtilityAddin
+  public class UtilityHandler : UtilityAddin
   {
     private bool _initialized = false;
 
-    public AddinUtilityHandler()
+    public UtilityHandler()
     {
       _initialized = true;
       Log.Debug("External SampleAssm Utility add-in: Initialized");
