+++ source/Xeno.ToolsHub.SampleAssmAddin/SampleAssmAddin.csproj
@@ -44,15 +44,15 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AddinAssembly.cs" />
-    <Compile Include="AddinOnStartHandler.cs" />
-    <Compile Include="AddinUtilityHandler.cs" />
-    <Compile Include="SampleAssmPreferenceHandler.cs" />
+    <Compile Include="Handlers\OnStartHandler.cs" />
+    <Compile Include="Handlers\UtilityHandler.cs" />
+    <Compile Include="Handlers\PreferenceHandler.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="SampleAssmPreferencesPage.cs">
+    <Compile Include="Views\PreferencesPage.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="SampleAssmPreferencesPage.Designer.cs">
-      <DependentUpon>SampleAssmPreferencesPage.cs</DependentUpon>
+    <Compile Include="Views\PreferencesPage.Designer.cs">
+      <DependentUpon>PreferencesPage.cs</DependentUpon>
     </Compile>
   </ItemGroup>
   <ItemGroup>
@@ -66,8 +66,8 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="SampleAssmPreferencesPage.resx">
-      <DependentUpon>SampleAssmPreferencesPage.cs</DependentUpon>
+    <EmbeddedResource Include="Views\PreferencesPage.resx">
+      <DependentUpon>PreferencesPage.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
