+++ source/Xeno.ToolsHub.SampleAssmAddin/Views/PreferencesPage.cs
@@ -1,7 +1,7 @@
 ﻿/* Copyright Xeno Innovations, Inc. 2018
  * Date:    2018-12-17
  * Author:  Damian Suess
- * File:    SampleAssmPreferencesPage.cs
+ * File:    PreferencesPage.cs
  * Description:
  *
  */
@@ -11,11 +11,11 @@ using System.Windows.Forms;
 using Xeno.ToolsHub.Config;
 using Xeno.ToolsHub.ExtensionModel.Preferences;
 
-namespace Xeno.ToolsHub.SampleAssmAddin
+namespace Xeno.ToolsHub.SampleAssmAddin.Views
 {
-  public partial class SampleAssmPreferencesPage : Form, IPreferencePageForm
+  public partial class PreferencesPage : Form, IPreferencePageForm
   {
-    public SampleAssmPreferencesPage()
+    public PreferencesPage()
     {
       InitializeComponent();
     }
