+++ source/Xeno.ToolsHub.SampleAssmAddin/Views/PreferencesPage.resx
