+++ source/Xeno.ToolsHub.SampleXmlAddin/Views/PerferencePage.resx
